+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 
 ## ChangeLog
 
+#### Version 0.9.0 (not yet released)
+- [feature:] Added `getClients` that returns a list of available email clients (Android)
+- [__change__]: Replace `isAvailable` through `hasClient` and `hasAccount`.
+- [__change__]: Plugin does not add any permissions by itself like `GET_ACCOUNTS` or `READ_EXTERNAL_STORAGE`.
+- [__change__]: `isAvailable` does not request for missing permission (`GET_ACCOUNTS`).
+- [__change__]: `hasPermission` takes 3 arguments now. The first one has to be a value of `cordova.plugins.email.permission`.
+- [__change__]: `requestPermission` takes 3 arguments now. The first one has to be a value of `cordova.plugins.email.permission`.
+- [__change__]: Remove support lib from being installed (Android).
+- [__change__]: Remove deprecated namespace `plugin.email`.
+- [__change__]: Remove deprecated support for `isHTML`.
+- [__change__]: Change default value for `isHtml` to `false`.
+- [enhancement]: Skip chooser if there's only the default app (#302)
+- [enhancement]: Improve chooser to only display email clients
+- [enhancement]: Add `from` to specify the sending email account (iOS11+)
+- [bugfix:] Do not open old email draft [fixes #303]
+
 #### Version 0.8.15 (08.02.2018)
 - Fix iOS not working if `app:` wasn't specified.
 - Fix `attachments:` to accept a string.